% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{area_transf}
\alias{area_transf}
\title{area_transf}
\usage{
area_transf(data)
}
\arguments{
\item{data}{dataframe with categorical variable "area_name" indicating the name of the area in which the restaurant is located (prefecture)}
}
\value{
data
}
\description{
area_transf
}
\examples{
ex <- data.frame(area_name = c("Fukuoka-ken Itoshima-shi Maebarunishi"))
area_transf(ex) \%>\% head()

}
