% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_fct.R
\name{basic_area_transf}
\alias{basic_area_transf}
\title{basic_area_transf}
\usage{
basic_area_transf(data)
}
\arguments{
\item{data}{dataframe with categorical variable "area_name" indicating the name of the area in which the restaurant is located (prefecture)}
}
\value{
data
}
\description{
basic_area_transf
}
\examples{
ex <- data.frame(area_name = c("Fukuoka-ken Itoshima-shi Maebarunishi"))
basic_area_transf(ex) \%>\% head()

}
