% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_fcts.R
\name{count}
\alias{count}
\title{count}
\usage{
count(df, time, daterange, date1, date2, day)
}
\arguments{
\item{df}{a dataframe you want to explore}

\item{time}{a type of timespan you want to explore - column of dataframe}

\item{daterange}{a vector of date you want to limit your exploration to - if needed}

\item{date1}{the minimum numeric corresponding to the later time you want to limit your exploration to - if needed}

\item{date2}{the maximum numeric corresponding to the later time you want to limit your exploration to - if needed}

\item{day}{a vector of day name you want to limit your exploration to - if needed}
}
\value{
the number of people visiting restaurant within the parameters selected
}
\description{
count
}
\examples{
# count(dataframe, visit_time,3,"2016-01-01","2016-01-03",c("monday","tuesday"))

}
