% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve_fct.R
\name{dn_transf}
\alias{dn_transf}
\title{dn_transf}
\usage{
dn_transf(data)
}
\arguments{
\item{data}{dataframe with categorical variable "visit_time" indicating time of visit in a given restaurant}
}
\value{
data
}
\description{
dn_transf
}
\examples{
ex <- data.frame(visit_time = c("13:00:00","02:00:00","15:00:00"))
dn_transf(ex) \%>\% head()

}
