% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_fct.R
\name{genre_transf}
\alias{genre_transf}
\title{genre_transf}
\usage{
genre_transf(data)
}
\arguments{
\item{data}{dataframe with categorical variable "genre_name" indicating style of restaurant which was booked/visited}
}
\value{
data
}
\description{
genre_transf
}
\examples{
ex <- data.frame(genre_name = c("Yakiniku/Korean food","Cafe/Sweets","Creative cuisine"))
genre_transf(ex) \%>\% head()

}
