% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{goldd_transf}
\alias{goldd_transf}
\title{goldd_transf}
\usage{
goldd_transf(data)
}
\arguments{
\item{data}{dataframe with categorical variable "visit_date" indicating date of visit}
}
\value{
data
}
\description{
goldd_transf
}
\examples{
ex <- data.frame(visit_date = c("2016-01-13","2016-04-02","2017-02-11"))
goldd_transf(ex) \%>\% head()

}
