% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_eval.R
\name{graph_dif}
\alias{graph_dif}
\title{graph_dif}
\usage{
graph_dif(test, n)
}
\arguments{
\item{test}{data frame containing a variable called y_true (the true value) and a variable called xgbpred (the prediction)}

\item{n}{the number of points we want to see on the evaluation plot}
}
\value{
a plot showing the difference between real and predicted value
}
\description{
graph_dif
}
\examples{
# graph_dif(evaluation_data_frame, 100)
}
