% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{lag_transf}
\alias{lag_transf}
\title{lag_transf}
\usage{
lag_transf(data)
}
\arguments{
\item{data}{dataframe with categorical variable "id" the identifier of each different restaurant in the data, and the quantitative variable "visitors" indicating the amount of people who visited}
}
\value{
data
}
\description{
lag_transf
}
\examples{
ex <- data.frame(id = "air1", visitors = 32)

}
