% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_fct.R
\name{lat_transf}
\alias{lat_transf}
\title{lat_transf}
\usage{
lat_transf(data)
}
\arguments{
\item{data}{dataframe with quantitative variable "longitude" indicating lon coordinates of a given restaurant}
}
\value{
data with one hot encoded additionnal variables for latitude coordinates of the locations of every id in our data
}
\description{
lat_transf
}
\examples{
ex <- data.frame(latitude = c(34.66474,34.75695,35.67492))
lat_transf(ex) \%>\% head()

}
