% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve_fct.R
\name{latency_transf}
\alias{latency_transf}
\title{latency_transf}
\usage{
latency_transf(data)
}
\arguments{
\item{data}{dataframe with categorical variable "reserve_date" indicating date on which reservation was made}
}
\value{
data
}
\description{
latency_transf
}
\examples{
ex <- data.frame(reserve_date = c("2016-01-13"), visit_date = c("2017-03-01"))
latency_transf(ex) \%>\% head()

}
