% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_fct.R
\name{lon_transf}
\alias{lon_transf}
\title{lat_transf}
\usage{
lon_transf(data)
}
\arguments{
\item{data}{dataframe with quantitative variable "latitude" indicating lat coordinates of a given restaurant}
}
\value{
data
}
\description{
lat_transf
}
\examples{
ex <- data.frame(longitude = c(139.7719,139.0363,139.6982))
lon_transf(ex) \%>\% head()

}
