% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{nth_day_trasnf}
\alias{nth_day_trasnf}
\title{nth_day_trasnf}
\usage{
nth_day_trasnf(data)
}
\arguments{
\item{data}{dataframe with categorical variable "visit_date" indicating date on which there was a visit in a given restaurant}
}
\value{
data
}
\description{
nth_day_trasnf
}
\examples{
date_info <- system.file("cleaned_csv","date_info.csv",package = "recruitjap")
ex <- data.frame(id = c("id1"), visit_date = c("2016-01-01"))
nth_day_trasnf(ex)

}
