% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_fcts.R
\name{personal_overview}
\alias{personal_overview}
\title{personal_overview}
\usage{
personal_overview(df, time, daterange, date1, date2, day, input_id)
}
\arguments{
\item{df}{a dataframe you want to explore}

\item{time}{a type of timespan you want to explore - column of dataframe}

\item{daterange}{a vector of date you want to limit your exploration to - if needed}

\item{date1}{the minimum numeric corresponding to the later time you want to limit your exploration to - if needed}

\item{date2}{the maximum numeric corresponding to the later time you want to limit your exploration to - if needed}

\item{day}{a vector of day name you want to limit your exploration to - if needed}

\item{input_id}{a string of character specifying the id of the restaurant}
}
\value{
a plot of the number of visitors depending on parameters
}
\description{
personal_overview
}
\examples{
# personal_overview(data,visit_time,3,"2016-01-01","2016-01-03",monday","1")

}
