% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_fcts.R
\name{predictive_table}
\alias{predictive_table}
\title{predictive_table}
\usage{
predictive_table(df, ids, date)
}
\arguments{
\item{df}{a data frame}

\item{ids}{a vector of string containing the id of restautant you want the prediction for}

\item{date}{a data factor with the date you want the prediction for}
}
\value{
a table containing the prediction of visitors for the ids and date selected
}
\description{
predictive_table
}
\examples{
# predictive_table(dataframe1, c("air243s679j","air11e4&"),"2016-11-11")
}
