% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve_fct.R
\name{rdayweek_transf}
\alias{rdayweek_transf}
\title{rdayweek_transf}
\usage{
rdayweek_transf(data)
}
\arguments{
\item{data}{dataframe with categorical variable "day_of_the_week_reserve" indicating the day on which reservation was made}
}
\value{
data
}
\description{
rdayweek_transf
}
\examples{
ex <- data.frame(day_of_the_week_reserve = c("Friday","Tuesday","Friday"))
rdayweek_transf(ex) \%>\% head()

}
