% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve_fct.R
\name{rholflg_transf}
\alias{rholflg_transf}
\title{rholflg_transf}
\usage{
rholflg_transf(data)
}
\arguments{
\item{data}{dataframe with binary variable "holiday_flg_reserve" indicating if corresponding date is a holiday in japan}
}
\value{
data
}
\description{
rholflg_transf
}
\examples{
ex <- data.frame(holiday_flg_reserve = c("1","0","1","0","0"))
rholflg_transf(ex) \%>\% head()

}
