% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_eval.R
\name{rmlse_fun}
\alias{rmlse_fun}
\title{rmlse_fun}
\usage{
rmlse_fun(prediction, true_value)
}
\arguments{
\item{prediction}{a data frame with the prediction}

\item{true_value}{a data frame with the true values}
}
\value{
the Root Mean Squared Logarithmic Error (RMSLE)
}
\description{
rmlse_fun
}
\examples{
# rmlse_fun(prediction_data_frame, true_value_data_frame)
}
