% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve_fct.R
\name{ryear_transf}
\alias{ryear_transf}
\title{ryear_transf}
\usage{
ryear_transf(data)
}
\arguments{
\item{data}{dataframe with categorical variable "reserve_date" indicating date of reservation}
}
\value{
data
}
\description{
ryear_transf
}
\examples{
ex <- data.frame(reserve_date = c("2016-01-13","2016-04-02","2017-02-11"))
ryear_transf(ex) \%>\% head()

}
