% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{same_features}
\alias{same_features}
\title{same_features}
\usage{
same_features(data, sample)
}
\arguments{
\item{data}{a data frame}

\item{sample}{a data frame}
}
\value{
a list containing the two identical (in terms of variabe/columns) data frames
}
\description{
same_features
}
\examples{
# same_features(data_frame1, data_frame2)

}
