% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{shinyapp}
\alias{shinyapp}
\title{shinyapp_jap}
\usage{
shinyapp()
}
\value{
shinyapp of our project
}
\description{
shinyapp_jap
}
\examples{
dir <- system.file("shiny","template_app1.R", package = "recruitjap")
# shinyapp(dir) don't run or it just won't stop the connexion!

}
