% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_fcts.R
\name{sortLvls.fnc}
\alias{sortLvls.fnc}
\title{sortLvls.fnc}
\usage{
sortLvls.fnc(oldFactor, levelOrder)
}
\arguments{
\item{oldFactor}{the factor column you want to sort arbitrarely - instead of alphabetically}

\item{levelOrder}{a numeric vector specifying the sorting you want}
}
\value{
levels of factor sorting the way you want
}
\description{
sortLvls.fnc
}
\examples{
# str <- as.factor(c("dimanche","samedi")) sortLvls.fnc(str,c(2,1))
}
