% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{test_train}
\alias{test_train}
\title{test_train}
\usage{
test_train(data)
}
\arguments{
\item{data}{a data set (must not be ordered) on which a test train split should be applied}
}
\value{
list containing split_index (the index at which the data was split), id_visit_date (a data frame with id and visit date),visitors (a data frame containing the visitors), train (the train data frame), test (the test data frame))
}
\description{
test_train
}
\examples{
# test_train(data_frame)

}
