% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_fct.R
\name{vday_transf}
\alias{vday_transf}
\title{vday_transf}
\usage{
vday_transf(data)
}
\arguments{
\item{data}{}
}
\value{
returns a one hot encoded dataframe for the day of the visit
}
\description{
vday_transf
}
\examples{
vday_transf(air_data)
}
