% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_fct.R
\name{vdayweek_transf}
\alias{vdayweek_transf}
\title{vdayweek_transf}
\usage{
vdayweek_transf(data)
}
\arguments{
\item{data}{dataframe with categorical variable "day_of_the_week_visit" indicating the day on which there was a visit}
}
\value{
data
}
\description{
vdayweek_transf
}
\examples{
ex <- data.frame(day_of_the_week_visit = c("Friday","Tuesday","Friday"))
vdayweek_transf(ex) \%>\% head()

}
