% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_fct.R
\name{vholflg_transf}
\alias{vholflg_transf}
\title{vholflg_transf}
\usage{
vholflg_transf(data)
}
\arguments{
\item{data}{dataframe with binary variable "holiday_flg_visit" indicating if date of visit was on a holiday in japan}
}
\value{
data
}
\description{
vholflg_transf
}
\examples{
ex <- data.frame(holiday_flg_visit = c("1","0","1","0","0"))
vholflg_transf(ex) \%>\% head()

}
