% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_fct.R
\name{vtime_transf}
\alias{vtime_transf}
\title{vtime_transf}
\usage{
vtime_transf(data)
}
\arguments{
\item{data}{dataframe with categorical variable "visit_time" indicating time of visit}
}
\value{
data
}
\description{
vtime_transf
}
\examples{
ex <- data.frame(visit_time = c("13:00:00","02:00:00","15:00:00"))
vtime_transf(ex) \%>\% head()

}
