% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{wealth_transf}
\alias{wealth_transf}
\title{wealth_transf}
\usage{
wealth_transf(data)
}
\arguments{
\item{data}{dataframe with categorical variable "area_name" indicating the name of the area in which the restaurant is located (prefecture)}
}
\value{
data
}
\description{
wealth_transf
}
\examples{
ex <- data.frame(area_name = c("Fukuoka-ken Itoshima-shi Maebarunishi"))
basic_area_transf(ex) \%>\% head()

}
