% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{wknd_transf}
\alias{wknd_transf}
\title{wknd_transf}
\usage{
wknd_transf(data)
}
\arguments{
\item{data}{dataframe with categorical variable "day_of_the_week_visit" indicating the day on which there was a visit for a given restaurant}
}
\value{
data
}
\description{
wknd_transf
}
\examples{
ex <- data.frame(day_of_the_week_visit = c("Friday","Tuesday","Friday"))
wknd_transf(ex) \%>\% head()

}
