% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_fcts.R
\name{xgboost_output}
\alias{xgboost_output}
\title{xgboost_output}
\usage{
xgboost_output(dTRAIN, eta, gamma, max_depth, min_child_weight, subsample,
  colsample_bytree)
}
\arguments{
\item{dTRAIN}{train matrix in xgb_matrix format}

\item{eta}{XGBoost parameter eta}

\item{gamma}{XGBoost parameter gamma}

\item{max_depth}{XGBoost parameter max_depth}

\item{min_child_weight}{XGBoost parameter min_child_weight}

\item{subsample}{XGBoost parameter subsample}

\item{colsample_bytree}{XGBoost parameter colsample_bytree}
}
\value{
a classifier, trained by xgboost
}
\description{
xgboost_output
}
\examples{
# xgboost_output(Train_matrix_xgb_matrix, 0.3, 0, 20, 1, 1, 1)
}
