% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_fcts.R
\name{xgboost_predict}
\alias{xgboost_predict}
\title{xgboost_predict}
\usage{
xgboost_predict(xgb1, dTEST)
}
\arguments{
\item{xgb1}{a classifier trained with xgboost}

\item{dTEST}{a test matrix in the xgbboost_matrix format}
}
\value{
returns the prediction
}
\description{
xgboost_predict
}
\examples{
#xgboost_predict(xgb_classifier, test_matrix_xgb_matrix)
}
